extract selectors and control bar from player component
