+++ packages/player/src/components/controls/play-toggle.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import {useSelector} from '@xstate/react'
 import cx from 'classnames'
 import {VideoContext} from '../../context/video-context'
-import {selectPaused} from '../player'
+import {selectPaused} from '../../selectors'
 
 export const PlayToggle: React.FC<any> = React.forwardRef<
   HTMLButtonElement,
