+++ packages/player/src/components/time-controls/current-time-display.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import {useSelector} from '@xstate/react'
 import cx from 'classnames'
-import {selectFormattedTime} from '../player'
 import {VideoContext} from '../../context/video-context'
+import {selectFormattedTime} from '../../selectors'
 
 export const CurrentTimeDisplay: React.FC<any> = ({className}) => {
   const {videoService} = React.useContext(VideoContext)
